Move cursor around properly when adding borders
**Description**
Allow the cursor to better handle its position relative to any borders
given. Also adds a `Point` struct for `Window` to make it a bit easier
to figure out where you are.

🚨**Testing:**
Start up a window and make sure that the typing begins in the right
place given a border

**Effects**
Closes #14